Conflicts:
	qtcreator.pri
	qtcreator.qbp
	src/libs/utils/utils.pro

+++ src/plugins/qnx/blackberryapplicationrunner.cpp
@@ -155,11 +155,15 @@ void BlackBerryApplicationRunner::startFinished(int exitCode, QProcess::ExitStat
         const QString errorString = (m_launchProcess->error() != QProcess::UnknownError)
                 ? m_launchProcess->errorString() : tr("Launching application failed");
         emit startFailed(errorString);
+        reset();
     }
 }
 
 ProjectExplorer::RunControl::StopResult BlackBerryApplicationRunner::stop()
 {
+    if (m_stopping)
+        return ProjectExplorer::RunControl::AsynchronousStop;
+
     m_stopping = true;
 
     QStringList args;
@@ -179,14 +183,13 @@ ProjectExplorer::RunControl::StopResult BlackBerryApplicationRunner::stop()
         m_stopProcess->setEnvironment(m_environment.toStringList());
     }
 
-
     m_stopProcess->start(m_deployCmd, args);
     return ProjectExplorer::RunControl::AsynchronousStop;
 }
 
 bool BlackBerryApplicationRunner::isRunning() const
 {
-    return m_running && !m_stopping;
+    return m_running;
 }
 
 qint64 BlackBerryApplicationRunner::pid() const
@@ -247,7 +250,7 @@ void BlackBerryApplicationRunner::tailApplicationLog()
 {
     // TODO: Reading the log using qconn instead?
 
-    if (m_tailProcess && m_tailProcess->isProcessRunning())
+    if (m_stopping || (m_tailProcess && m_tailProcess->isProcessRunning()))
         return;
 
     QTC_CHECK(!m_appId.isEmpty());
@@ -344,9 +347,11 @@ void BlackBerryApplicationRunner::reset()
     m_stopping = false;
 
     m_runningStateTimer->stop();
-    m_runningStateProcess->terminate();
-    if (!m_runningStateProcess->waitForFinished(1000))
-        m_runningStateProcess->kill();
+    if (m_runningStateProcess) {
+        m_runningStateProcess->terminate();
+        if (!m_runningStateProcess->waitForFinished(1000))
+            m_runningStateProcess->kill();
+    }
 
     if (m_tailProcess && m_tailProcess->isProcessRunning())
         killTailProcess();
